--- config/modules.config.js
@@ -1,59 +1,0 @@
-module.exports = {
-  core: [
-    'arcade/arcade.js',
-    'archive/archive.js',
-    'blackjack/blackjack.js',
-    'chat/chat.js',
-    'chess/chess.js',
-    'devcenter/devcenter.js',
-    'encrypt/encrypt.js',
-    'explorer/explorer.js',
-    'myqrcode/myqrcode.js',
-    'poker/poker.js',
-    'post/post.js',
-    'qrscanner/qrscanner.js',
-    'relay/relay.js',
-    'registry/registry.js',
-    'saitolicious/saitolicious.js',
-    'settings/settings.js',
-    'invites/invites.js',
-    // 'staking/staking.js',
-    'settlers/settlers.js',
-    'solitrio/solitrio.js',
-    'wordblocks/wordblocks.js',
-    'calendar/calendar.js',
-    'website/website.js',
-    'redsquare/redsquare',
-    'stun/stun',
-    'video/video'
-  ],
-  lite: [
-    'arcade/arcade.js',
-    'archive/archive.js',
-    'blackjack/blackjack.js',
-    'chess/chess.js',
-    'chat/chat.js',
-    'debug/debug.js',
-    'devcenter/devcenter.js',
-    'invites/invites.js',
-    'encrypt/encrypt.js',
-    'myqrcode/myqrcode.js',
-    'qrscanner/qrscanner.js',
-    'poker/poker.js',
-    'post/post.js',
-    'relay/relay.js',
-    'registry/registry.js',
-    'saitolicious/saitolicious.js',
-    'settings/settings.js',
-    // 'staking/staking.js',
-    'email/email.js',
-    'settlers/settlers.js',
-    'solitrio/solitrio.js',
-    'wordblocks/wordblocks.js',
-    'calendar/calendar.js',
-    'website/website.js',
-    'redsquare/redsquare',
-    'stun/stun',
-    'video/video'
-  ]
-}
