+++ lib/saito/ui/saito-overlay/saito-overlay.js
@@ -32,9 +32,7 @@ class SaitoOverlay {
    * @param app - the Saito Application
    * @param mod - the calling module
    */
-  attachEvents(app, mod, callback) {
-    callback()
-  }
+  attachEvents(app, mod) {}
 
   /**
    * Renders the Overlay with the given html and attaches events to close it
@@ -44,9 +42,8 @@ class SaitoOverlay {
    * @param mycallback - a function to run when the user closes the overlay
    *
    */
-  show(app, mod, html, mycallback = null, eventCallback = null) {
+  show(app, mod, html, mycallback = null) {
     this.render(app, mod);
-
 
     let overlay_self = this;
 
@@ -71,10 +68,8 @@ class SaitoOverlay {
       }
 
       overlay_backdrop_el.onclick = (e) => {
-
         overlay_self.hide(mycallback);
       };
-      this.attachEvents(app, mod, eventCallback);
     } catch (err) {
       console.error("OVERLAY ERROR:", err);
     }
