+++ mods/devcenter/web/style.css
@@ -8,20 +8,16 @@
   padding-left: 30px;
   padding-bottom: 30px;
 }
-
 .welcome-link {
   color: var(--saito-red);
   font-weight: bold;
 }
-
 .welcome-link a {
   text-decoration: none;
 }
-
 .welcome-text {
   font-size: 0.9em;
 }
-
 h3 {
   margin: 0px;
 }
@@ -36,7 +32,6 @@ hr {
   color: white;
   font-weight: bold;
 }
-
 .email-apps-item.active {
   background-color: var(--saito-red);
   color: white;
@@ -49,7 +44,6 @@ hr {
   max-width: 800px;
   padding: 0 2em;
 }
-
 .welcome-container p {
   margin: 0 0 20px;
 }
@@ -138,7 +132,7 @@ hr {
   transition: all 0.25s ease-in-out;
 }
 
-.learn>div {
+.learn > div {
   margin: 0 0 1em 1em;
 }
 
@@ -158,6 +152,7 @@ hr {
 
 /*  Stun UI CSS */
 .stun-container {
+  padding: 1.5rem;
   font-size: 1.2rem;
   transition: all 0.3s;
 }
@@ -172,7 +167,6 @@ hr {
   border: 1px solid black;
   width: 80%;
 }
-
 .stun-video::after {
   content: "Local Stream";
   display: block;
@@ -189,7 +183,6 @@ hr {
   display: flex;
   /* justify-content: center; */
 }
-
 .stun-container__bar button {
   color: #f71f3d;
   border: 1px solid #f71f3d;
@@ -204,19 +197,19 @@ hr {
   background-color: #f71f3d;
   color: white;
 }
-
 button.button-active {
   background-color: #f71f3d;
   color: white;
 }
 
-.stun-information {}
+.stun-information {
+  margin-top: 6rem;
+}
 
 .stun-information .name {
   font-weight: bold;
   font-size: 1.2rem;
 }
-
 .stun-information .data {
   font-size: 1.2rem;
   color: rgba(0, 0, 0, 0.541);
@@ -242,18 +235,16 @@ button.button-active {
 .listeners-container .input-container {
   margin-bottom: 3rem;
 }
-
-.listeners-container>label {
+.listeners-container > label {
   font-weight: bold;
 }
-
-.listeners-container .input-container>textarea {
+.listeners-container .input-container > textarea {
   border-radius: 8px;
   width: 100%;
   margin-top: 0.5rem;
 }
 
-.listeners-container .input-container>button {
+.listeners-container .input-container > button {
   right: 0;
 }
 
@@ -265,7 +256,6 @@ button.button-active {
   from {
     opacity: 0;
   }
-
   to {
     opacity: 1;
   }
@@ -275,33 +265,26 @@ button.button-active {
   input[type="checkbox"] {
     margin: 0 0.65em;
   }
-
   .main {
     max-width: unset;
     padding: 1em 0;
   }
-
   .email-body {
     padding: 3px;
   }
-
   .email-header {
     padding: 0 0.4em 0 0.4em;
   }
-
   .email-main {
     padding-right: unset;
   }
-
   .email-container {
     display: block;
     padding-left: 10px;
   }
-
   .email-sidebar {
     display: none;
   }
-
   .email-bars-menu {
     position: absolute;
     z-index: 10000;
@@ -312,11 +295,9 @@ button.button-active {
     display: block;
     box-shadow: 2px 2px 5px 0px rgba(0, 0, 0, 0.75);
   }
-
   .email-apps {
     border-bottom: 1px solid black;
   }
-
   #qrcode img {
     width: 66%;
     margin: 0.5em 16.6% 1em;
@@ -324,7 +305,6 @@ button.button-active {
 }
 
 /*  video chat */
-
 /* video chat */
 .body .button {
   cursor: pointer;
@@ -336,7 +316,6 @@ button.button-active {
   padding: 15px;
   font-size: 2rem;
 }
-
 .body {
   background: linear-gradient(96.79deg, #f30c4f 0%, #f8c062 91.87%) !important;
   border: 1px solid black;
@@ -345,15 +324,12 @@ button.button-active {
   position: absolute !important;
   top: 5rem;
 }
-
 .body section {
   background: linear-gradient(96.79deg, #f30c4f 0%, #f8c062 91.87%);
 }
-
 .body .main {
   background: linear-gradient(96.79deg, #f30c4f 0%, #f8c062 91.87%);
 }
-
 .body .saito-logo {
   position: absolute;
   width: 80px;
@@ -371,7 +347,6 @@ button.button-active {
   transform: matrix(1, 0, 0, -1, 0, 0);
   background: none;
 }
-
 .body .disconnect_btn {
 
   /* left: 83.96%;
@@ -412,12 +387,11 @@ button.button-active {
   border-radius: 1.92459px;
   background: white;
 
-  width: 100%;
-  height: 100%;
+ width: 100%;
+ height: 100%;
   object-fit: cover;
-
+  
 }
-
 .body .box1 {
   /* grid-row: 1/3; */
 }
@@ -433,88 +407,31 @@ button.button-active {
   cursor: pointer;
 }
 
-.body .footer {
+.body  .footer {
   margin-top: 20px;
   display: flex;
-
+  justify-content: space-between;
+ 
   padding: 20px;
-  background: radial-gradient(100.52% 47851.24% at 0% 0%,
+  background: radial-gradient(
+      100.52% 47851.24% at 0% 0%,
       #f66659 0%,
       #f21d3b 51.56%,
-      #f66659 100%)
-    /* warning: gradient uses a rotation that is not supported by CSS and may not behave as expected */
-  ;
+      #f66659 100%
+    )
+    /* warning: gradient uses a rotation that is not supported by CSS and may not behave as expected */;
   transform: matrix(1, 0, 0, -1, 0, 0);
   color: white;
   align-items: center;
 }
 
-
-.disconnect_btn {
-  margin-left: auto;
-}
 
 
 .footer .fa {
   color: white;
-  font-size: 2rem;
+  font-size: 2.7rem;
   cursor: pointer;
   transform: matrix(1, 0, 0, -1, 0, 0);
-  margin-right: 3rem;
-}
-
-/* video chat */
-
-
-
-/*  Stun */
-
-
-.my-stun-container h2 {
-  margin-bottom: .1rem;
-}
-
-.my-stun-container>p {
-  margin-bottom: 1rem;
-}
-
-.my-stun-container-actions {
-  margin-bottom: 1rem;
-}
-
-
-
-.my-stun-container-join p {
-  font-size: 1rem;
-  font-weight: bold;
 
 }
 
-.my-stun-container-create {
-  margin-bottom: 1rem;
-}
-
-.my-stun-container-create p {
-  font-size: 1rem;
-  font-weight: bold;
-}
-
-
-.my-stun-container-join {}
-
-.my-stun-container-join>div {
-  display: grid;
-  grid-template-columns: 13rem 13rem;
-  align-items: center;
-}
-
-.my-stun-container-join input {
-  margin-right: 1rem;
-  padding-left: 1rem;
-  height: 50.798px;
-
-}
-
-.my-stun-container-join input:focus {
-  outline: none;
-}
