+++ mods/stun/lib/email-appspace/email-appspace.js
@@ -1,13 +1,13 @@
 //main
-const StunMainContainer = require("./email-appspace.template");
+const StunMainContainer = require("./../email-appspace/email-appspace.template");
 
 // components
-const StunComponentMyStun = require("../components/my-stun");
-const StunComponentListeners = require("../components/listeners");
-const StunComponentPeers = require("../components/peers");
+const StunComponentMyStun = require("./../components/my-stun");
+const StunComponentListeners = require("./../components/listeners");
+const StunComponentPeers = require("./../components/peers");
 
-const { vanillaToast } = require("vanilla-toast");
-const VideoChat = require('../../../../lib/saito/ui/video-chat/video-chat');
+const {vanillaToast} = require("vanilla-toast");
+const VideoChat = require('./../../../../lib/saito/ui/video-chat/video-chat');
 
 
 class Container {
@@ -43,10 +43,10 @@ class Container {
     }
 
 
-    render(app = this.app, mod = this.mod) {
-        if (!document.querySelector('.stun-container'))
-            document.querySelector('#email-appspace').innerHTML = sanitize(StunMainContainer(app, mod));
-
+    render(app = this.app , mod = this.mod) {
+        if (!document.querySelector('.stun-container')) 
+            document.querySelector('.email-container').innerHTML = sanitize(StunMainContainer(app, mod));
+       
         // render the selected tab
         const Tab = this.mapTabToTemplate[this.selectedTab];
         Tab.render(app, mod);
@@ -67,7 +67,6 @@ class Container {
         const generate_button = document.querySelector('#generate');
         const update_button = document.querySelector('#update');
         const input = document.querySelector('#input_address');
-        let self = this;
 
         app.connection.on('stun-update', (app) => {
             console.log('stun update', app);
@@ -86,7 +85,7 @@ class Container {
 
 
             // create new RTC connection 
-            const createPeerConnection = async () => {
+            const createPeerConnection = async() => {
                 const stun_mod = app.modules.returnModule('Stun');
                 let reply = {
                     answer: "",
@@ -303,19 +302,21 @@ class Container {
         // event listeners
 
         // change selected tab
-        $('.menu').on('click', function (e) {
+        $('.menu').on('click', function(e) {
 
             if ($(this).attr('data-id') === this.selectedTab) return;
             // add ui class to button
             $('.menu').removeClass('button-active');
             $(this).addClass('button-active');
-            self.selectedTab = $(this).attr('data-id');
+            this.selectedTab = $(this).attr('data-id');
 
-            self.render(app, mod);
+            this.render(app, mod);
+
+
         });
 
         //connect with peer
-        $(".stun-container").on('click', '#connectTo', function (e) {
+        $(".stun-container").on('click', '#connectTo', function(e) {
 
 
 
@@ -359,7 +360,7 @@ class Container {
 
             const createPeerConnection = new Promise((resolve, reject) => {
                 const stun_mod = app.modules.returnModule('Stun');
-                const execute = async () => {
+                const execute = async() => {
 
                     try {
                         const pc = new RTCPeerConnection({
