+++ mods/stun/lib/email-appspace/email-appspace.template.js
@@ -3,7 +3,10 @@ module.exports = (app, mod) => {
   let html = `
      <div class="stun-container">
      <div class="stun-container__bar container">
-    
+       <button id="my-stun-btn" data-id="my-stun" class="menu button-active">
+         Connect
+       </button>
+     
      
      </div>
