+++ mods/video/video.js
@@ -31,12 +31,12 @@ class Video extends ModTemplate {
 
 
     respondTo(type) {
-        // if (type == 'email-appspace') {
-        //     let obj = {};
-        //     obj.render = this.renderEmail;
-        //     obj.attachEvents = this.attachEmailEvents;
-        //     return obj;
-        // }
+        if (type == 'email-appspace') {
+            let obj = {};
+            obj.render = this.renderEmail;
+            obj.attachEvents = this.attachEmailEvents;
+            return obj;
+        }
         return null;
     }
 
@@ -66,6 +66,7 @@ class Video extends ModTemplate {
 
                 let my_pubkey = app.wallet.returnPublicKey();
                 if (tx.msg.answer) {
+
                     if (my_pubkey === tx.msg.answer.offer_creator) {
                         if (app.BROWSER !== 1) return;
                         console.log("current instance: ", my_pubkey, " answer room: ", tx.msg.answer);
@@ -87,6 +88,8 @@ class Video extends ModTemplate {
                         } else {
                             console.log("peer connection not found");
                         }
+
+
                     }
                 }
                 if (tx.msg.rooms) {
@@ -98,11 +101,14 @@ class Video extends ModTemplate {
                     if (app.BROWSER !== 1) return;
 
                     const offer_creator = tx.msg.offers.offer_creator;
+
                     // offer creator should not respond
                     if (my_pubkey === offer_creator) return;
-                    console.log("offers received from ", tx.msg.offers.offer_creator);
+                    console.log("offers received from ", tx.msg.offers.offer_creator, tx.msg.offers);
+
                     // check if current instance is a recipent
                     const index = tx.msg.offers.offers.findIndex(offer => offer.recipient === my_pubkey);
+
                     if (index !== -1) {
                         video_self.acceptOfferAndBroadcastAnswer(app, offer_creator, tx.msg.offers.offers[index]);
                     }
@@ -183,6 +189,7 @@ class Video extends ModTemplate {
         // lite clients are not allowed to run this
         if (this.app.BROWSER === 0) {
             let newtx2 = this.app.wallet.createUnsignedTransaction();
+
             // newtx2.transaction.to.push(new saito.default.slip(this.app.network.peers[this.app.network.peers.length - 1].returnPublicKey()));
 
             // console.log('sending to ', this.app.network.peers[this.app.network.peers.length - 1].returnPublicKey(), this.rooms);
@@ -224,12 +231,21 @@ class Video extends ModTemplate {
                 pc.onicecandidate = (ice) => {
                     if (!ice || !ice.candidate || !ice.candidate.candidate) {
                         console.log('ice candidate check closed');
+
                         let video_mod = app.modules.returnModule("Video");
                         video_mod.peer_connections[offer_creator] = pc;
+
                         video_mod.broadcastAnswer(video_mod.app.wallet.returnPublicKey(), offer_creator, reply);
                         return;
+
                     };
+
                     reply.ice_candidates.push(ice.candidate);
+
+
+
+
+
                 }
 
                 pc.onconnectionstatechange = e => {
@@ -270,7 +286,7 @@ class Video extends ModTemplate {
                 const localStream = await navigator.mediaDevices.getUserMedia({ video: true, audio: true });
                 localStream.getTracks().forEach(track => {
                     pc.addTrack(track, localStream);
-                    console.log('got my local stream');
+                    console.log('got local stream for answerer');
                 });
 
                 let video_self = app.modules.returnModule("Video");
@@ -282,7 +298,7 @@ class Video extends ModTemplate {
                 const remoteStream = new MediaStream();
                 pc.addEventListener('track', (event) => {
                     let video_self = app.modules.returnModule("Video");
-                    console.log('got remote stream from offer creator ', event.streams);
+                    console.log('got remote stream ', event.streams);
                     event.streams[0].getTracks().forEach(track => {
                         remoteStream.addTrack(track);
                     });
@@ -339,9 +355,7 @@ class Video extends ModTemplate {
         const html = `
         <div style="background-color: white; padding: 2rem 3rem; border-radius: 8px; display:flex; flex-direction: column; align-items: center; justify-content: center; align-items:center">
            <p style="font-weight: bold; margin-bottom: 3px;">  Invite Code: </p>
-           <div style="display: flex; align-item: center;"> 
-           <div style="margin-right: .5rem" id="copyVideoInviteCode"> <i class="fa fa-copy"> </i> </div> <p style="margin-right: .5rem"> ${roomCode} </p> 
-           </div>
+           <p> ${roomCode} </p>
         </div>
         `
 
@@ -374,17 +388,11 @@ class Video extends ModTemplate {
 
         const overlay = new SaitoOverlay(this.app);
 
-        overlay.show(this.app, video_self, html, null, () => {
-            console.log("attaching copy event")
-            document.querySelector('#copyVideoInviteCode i').addEventListener('click', (e) => {
-                navigator.clipboard.writeText(roomCode);
-                document.querySelector("#copyVideoInviteCode").textContent = "Copied to clipboard";
-            });
-        });
+        overlay.show(this.app, video_self, html);
+
+
 
     }
-
-
 
     createPeerConnectionOffer(publicKey) {
         const stun_mod = this.app.modules.returnModule('Stun');
@@ -407,6 +415,7 @@ class Video extends ModTemplate {
 
                             let offer_sdp = pc.localDescription;
                             resolve({ recipient: publicKey, offer_sdp, ice_candidates, pc });
+                            // stun_mod.broadcastIceCandidates(my_key, peer_key, ['savior']);
 
                             return;
                         } else {
@@ -594,7 +603,12 @@ class Video extends ModTemplate {
                 this.broadcastOffers(this.app.wallet.returnPublicKey(), offers);
             } else {
                 const localStream = await navigator.mediaDevices.getUserMedia({ video: true, audio: true });
+                localStream.getTracks().forEach(track => {
+                    console.log('got local stream for answerer');
+                });
+
                 let video_self = this.app.modules.returnModule("Video");
+
                 video_self.videoChat.show(new RTCPeerConnection({}));
                 video_self.videoChat.addLocalStream(localStream);
                 console.log("you are the only participant in the room");
@@ -644,9 +658,12 @@ class Video extends ModTemplate {
 
     broadcastOffers(offer_creator, offers) {
         let newtx = this.app.wallet.createUnsignedTransaction();
+
+
         console.log('broadcasting offers');
         for (let i = 0; i < offers.length; i++) {
             newtx.transaction.to.push(new saito.default.slip(offers[i].recipient));
+
         }
 
         newtx.msg.module = "Video";
